requirements.txt
pikepdf==7.2.0
Pillow==9.5.0
